+++ packages/core/button/src/components/Button.js
@@ -40,8 +40,7 @@ const createStyledComponent = () => {
   // Override pseudo-state specificity.
   // This is necessary because we don't know what DOM element the custom component will render.
   const component = styled(
-    //CustomComponentProxy is absolutely valid here, so this seems a
-    // problem with styled-components flow definitions
+    /* CustomComponentProxy is absolutely valid here, so this seems a problem with styled-components flow definitions */
     // $FlowFixMe
     CustomComponentProxy,
   )`&,a&,&:hover,&:active,&:focus{${getButtonStyles}}`;
@@ -161,7 +160,7 @@ class Button extends Component<ButtonProps, State> {
       isSelected,
       isDisabled,
     } = this.props;
-
+    // $FlowFixMe - Cannot call `getButtonProps` with `this` bound to `component` because `Button` [1] is incompatible with `Button` [2].
     const buttonProps = getButtonProps(this);
     const StyledComponent = this.getStyledComponent();