+++ packages/core/icon/glyph/select-clear.js
@@ -17,6 +17,6 @@ var _index2 = _interopRequireDefault(_index);
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
 var SelectClearIcon = function SelectClearIcon(props) {
-  return _react2.default.createElement(_index2.default, _extends({ dangerouslySetGlyph: '<svg width="24" height="24" viewBox="0 0 24 24" focusable="false" role="presentation"><path d="M12 21a9 9 0 1 1 0-18 9 9 0 0 1 0 18zm4.155-6.507L13.662 12l2.493-2.493a1.174 1.174 0 1 0-1.662-1.662L12 10.338 9.506 7.845a1.174 1.174 0 1 0-1.662 1.662L10.337 12l-2.493 2.493a1.176 1.176 0 0 0 1.662 1.663L12 13.662l2.493 2.494a1.172 1.172 0 0 0 1.662 0c.46-.46.46-1.203 0-1.663z" fill="#A5ADBA" fill-rule="evenodd"/></svg>' }, props));
+  return _react2.default.createElement(_index2.default, _extends({ dangerouslySetGlyph: '<svg width="24" height="24" viewBox="0 0 24 24" focusable="false" role="presentation"><g fill-rule="evenodd"><circle fill="currentColor" cx="12" cy="12" r="9"/><path d="M16.155 14.493a1.174 1.174 0 1 1-1.662 1.663L12 13.662l-2.494 2.494a1.172 1.172 0 0 1-1.662 0 1.176 1.176 0 0 1 0-1.663L10.337 12 7.844 9.507a1.176 1.176 0 0 1 1.662-1.662L12 10.338l2.493-2.493a1.174 1.174 0 1 1 1.662 1.662L13.662 12l2.493 2.493z" fill="inherit"/></g></svg>' }, props));
 };
 exports.default = SelectClearIcon;