+++ packages/core/navigation/src/components/js/NavigationWithTheme.js
@@ -4,7 +4,6 @@ import React from 'react';
 import { withTheme, ThemeProvider } from 'styled-components';
 import Navigation from './Navigation';
 
-// $FlowFixMe
 const NavigationWithTheme = withTheme(Navigation);
 
 const emptyTheme = {};