+++ packages/editor/editor-core/__tests__/plugins/gap-cursor/index.ts
@@ -5,9 +5,6 @@ import {
   p,
   blockquote,
   code_block,
-  ul,
-  li,
-  ol,
   panel,
   table,
   tr,
@@ -37,7 +34,6 @@ import gapCursorPlugin, {
 } from '../../../src/plugins/gap-cursor';
 import { pluginKey } from '../../../src/plugins/gap-cursor/pm-plugins/main';
 import codeBlockPlugin from '../../../src/plugins/code-block';
-import listsPlugin from '../../../src/plugins/lists';
 import rulePlugin from '../../../src/plugins/rule';
 import panelPlugin from '../../../src/plugins/panel';
 import tasksAndDecisionsPlugin from '../../../src/plugins/tasks-and-decisions';
@@ -58,7 +54,6 @@ describe('gap-cursor', () => {
         tablesPlugin,
         tasksAndDecisionsPlugin,
         codeBlockPlugin(),
-        listsPlugin,
         rulePlugin,
         panelPlugin,
       ],
@@ -73,8 +68,6 @@ describe('gap-cursor', () => {
 
   const blockNodes = {
     code_block: code_block({ language: 'java' })('{<>}'),
-    ul: ul(li(p('{<>}'))),
-    ol: ol(li(p('{<>}'))),
     panel: panel()(p('{<>}')),
     table: table()(tr(tdCursor)),
     decisionList: decisionList({ localId: 'test' })(