+++ packages/editor/editor-core/__tests__/plugins/type-ahead/pm-plugins/input-rules.ts
@@ -5,31 +5,7 @@ import {
   typeAheadQuery,
   insertText,
 } from '@atlaskit/editor-test-helpers';
-
-const createTypeAheadPlugin = ({
-  getItems,
-  selectItem,
-}: {
-  getItems?: Function;
-  selectItem?: Function;
-} = {}) => {
-  return {
-    pluginsOptions: {
-      typeAhead: {
-        trigger: '/',
-        getItems:
-          getItems !== undefined
-            ? getItems
-            : () => [{ title: '1' }, { title: '2' }, { title: '3' }],
-        selectItem:
-          selectItem !== undefined
-            ? selectItem
-            : (state, item, replaceWith) =>
-                replaceWith(state.schema.text(`${item.title} selected`)),
-      },
-    },
-  };
-};
+import { createTypeAheadPlugin } from './_create-type-ahead-plugin';
 
 describe('typeAhead input rules', () => {
   it('should convert trigger to a typeAheadQuery', () => {