+++ website/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/website
 
+## 2.1.5
+- [patch] Updated dependencies [eee2d45](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/eee2d45)
+  - @atlaskit/code@6.0.0
+  - @atlaskit/logo@8.1.3
+
 ## 2.1.4
 - [patch] Updated dependencies [c57e9c1](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/c57e9c1)
   - @atlaskit/media-test-helpers@14.0.4